// SDKs/velopayments.com/2.26.127/typescript/src/sdk/models/shared/payoraddressv2.ts
import { SpeakeasyMetadata, SpeakeasyBase } from "../../../internal/utils";



export class PayorAddressV2 extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=city" })
  city: string;

  @SpeakeasyMetadata({ data: "json, name=country" })
  country: string;

  @SpeakeasyMetadata({ data: "json, name=countyOrProvince" })
  countyOrProvince?: string;

  @SpeakeasyMetadata({ data: "json, name=line1" })
  line1: string;

  @SpeakeasyMetadata({ data: "json, name=line2" })
  line2?: string;

  @SpeakeasyMetadata({ data: "json, name=line3" })
  line3?: string;

  @SpeakeasyMetadata({ data: "json, name=line4" })
  line4?: string;

  @SpeakeasyMetadata({ data: "json, name=zipOrPostcode" })
  zipOrPostcode?: string;
}